+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 [1.1.0] 08.04.2026
 
+- [FAM-12] Transactions edit screen
+- [FAM-11] Transactions options buttons
 - [FAM-7] Transactions screen
 - [FAM-6] Wallets in dashboard
 - [FAM-5] Dashboard